+++ sqlite3/test/wasm/utils.dart
@@ -1,4 +1,3 @@
-import 'package:http/http.dart' as http;
 import 'package:sqlite3/wasm.dart';
 import 'package:test/scaffolding.dart';
 
@@ -9,11 +8,5 @@ Future<WasmSqlite3> loadSqlite3([SqliteEnvironment? environment]) async {
   final sqliteWasm =
       Uri.parse('http://localhost:$port/example/web/sqlite3.wasm');
 
-  final response = await http.get(sqliteWasm);
-  if (response.statusCode != 200) {
-    throw StateError(
-        'Could not load module (${response.statusCode} ${response.body})');
-  }
-
-  return WasmSqlite3.load(response.bodyBytes, environment);
+  return WasmSqlite3.loadFromUrl(sqliteWasm, environment: environment);
 }